+++ app/src/main/java/org/android/go/sopt/home/HomeActivity.kt
@@ -8,6 +8,7 @@ import org.android.go.sopt.R
 import org.android.go.sopt.databinding.ActivityHomeBinding
 import org.android.go.sopt.databinding.FragmentRepoBinding
 import org.android.go.sopt.home.home.HomeFragment
+import org.android.go.sopt.home.repo.RepoFragment
 
 class HomeActivity : AppCompatActivity() {
     private lateinit var binding: ActivityHomeBinding
